refactor(layout): extract AppHeader to its own component file
